python-fastapi-service/fastApiProject/requirements.txt
fastapi
uvicorn
langchain_openai
langchain
langchain_community
unstructured
python-dotenv
